Config file and topology
Inputs are now passed via a config.json file.
defaults.json is commited to the repository and contains the defaults
anything in config.json will overwrite the default.
The specific name of the file can be passed from the command line.

Added islands.py to house some functionalities on getting possible
islands on the feeder for anti-islanding/DTT etc. considerations.
TODO: capacities are now frozen with the network json, these should
somehow update with the changes/analysis.