Improved voltage monitoring and component flows
Voltage is now "sample" monitored throughout the feeder using a BFS.
This way:
    a) voltage results are available even if no PV is placed
    b) voltage results will always span the full distance of the feeder
Note: it has been noticed that both the minimum voltage and voltage
difference appear to violate the normal limits (0.95 and 3%
respectively) in the base 9500 feeder model.

The maximum, minimum, and minimum-absolute-value, of flows out of all
potential islands are calculated. The idea is that:
    * if the flows are always in the same direction (always exiting,
    or always entering) and they are above some threshold, then
    unintentional islanding is unlikely
    * if the flows change direction, then extra attention is paid, to
    the minimum-absolute-value. If this is small it means there is
    likely a moment in the simulation where the balance is close to 0
    which could lead to unintentional islanding.