Updates to transformer modeling
Updates to get things working with seconds and to
plot using plotly, enabling zooming in on second
level data.

Because the second level data is so intensive, the
input csv file now allows to specify which portion
should be plotted. This way, the run up day and
tail day don't have to be plotted along with the
incident of interest.