Adding presolve edits to run_opendss command
To model a DER at all hours of the day and be able to assign it
Inverter Controls it has to be modeled as a pvsystem (or storage).
The PV systems however, all get assigned a particular curve that
is always going to be zero overnight.
The new resource therefore, needs to get a new curve (flat)

This sort of thing can be achieved via the presolve edits input
which passes a list of commands to the dss engine right before solving
(and therefore after any batch edits etc.)

There are still issues with assigning different inverter controls the
way things are currently set up.
two functions for managing the controlled DER lists are added, however,
these are not worked into the full implementation yet.