Capability to apply shape to regulator and storage
For regulators csv files with the desired taps are provided.
This should usually be coupled with deactivating the associated
regulator control via either disable_all or disable_list.
The mapping from transformer object to csv file is done via
regulator_shape.

For storage elements, it is assumed that some change has been
done to the model to add loads in place of the storage elements.
That is, disable the storage and add similar load.
This addition can be done via change_lines_init.
Similarly, the new changes need to define a load shape that the
storage(i.e. load) should follow.
The mapping between the load name and the respective shape is
communicated via parameter "storage_shape"