Voltage dependent convergence flag
the open dss interface now takes an optional min_converged_vpu
argument. If the minimum voltage (last snapshot) is below this value
the solution is treated as not converged.

This is because OpenDSS seems to at times yield results with voltage
near zero.
For some reason, these are not always picked up in the voltage demand
interval, at least not when a single snapshot is calculated.