Adding bound search strategy to HCA search
When either an upper or a lower bound is not yet known during the
bisection search there are now two strategies: mult and add.
mult has been the strategy up until now, where the value was
doubled/halved until a bound was found.

With add a fixed value can be added/subtracted instead.
This is helpful when a fairly good starting point is known and it is
not desired to jump very far away from it during the search.